Move parts to own files
